+++ model.go
@@ -220,6 +220,17 @@ type EventStats struct {
 	MetaValue              string   `db:"meta_value" json:"meta_value"`
 }
 
+// EventListStats is the result type for a custom event list.
+type EventListStats struct {
+	Name     string            `db:"event_name" json:"name"`
+	Meta     map[string]string `db:"-" json:"meta"`
+	Visitors int               `json:"visitors"`
+	Count    int               `json:"count"`
+
+	// TODO optimize once maps are supported in the driver
+	Metadata [][]interface{} `db:"meta" json:"-"`
+}
+
 // ReferrerStats is the result type for referrer statistics.
 type ReferrerStats struct {
 	Referrer         string  `json:"referrer"`